refactored main => process_file, temp => reconstruct_file; added CLI support for subdivision levels
