+++ src/process_file.js
@@ -7,20 +7,22 @@ const dumpResult = require('./core/dumper');
 const loadFile = require('./core/file_loader');
 const parser = require('./core/obj_parser');
 
-function main(args) {
+function process_file(args) {
   let filename = args.input;
+  console.log(`loading ${filename}`);
   console.dir(args.input);
   const text = loadFile(filename);
   const mesh = parser.parseModel(text);
   const wrapper = new ShrinkWrapper(mesh, text);
 
-  wrapper.modify(5);
+  console.log(`processing with ${args.levels} subdivision levels`);
+  wrapper.modify(args.levels);
   dumpResult(wrapper, filename);
 
   console.log(`Processed ${wrapper.output.length} vertices.`);
 }
 
-const wrapped = performance.timerify(main);
+const wrapped = performance.timerify(process_file);
 
 const obs = new PerformanceObserver((list) => {
   console.log(list.getEntries()[0].duration);
@@ -29,4 +31,3 @@ const obs = new PerformanceObserver((list) => {
 obs.observe({entryTypes: ['function']});
 
 module.exports = wrapped;
-
